Former-commit-id: 122debd3717a088efec9b863764b3cbff5d6651c

+++ BackEnd/SpringBootServer/src/main/java/com/ssafy/backend/domain/administration/dto/info/OpenedStoreAdministrationTopFiveInfo.java
@@ -6,6 +6,6 @@ import lombok.Builder;
 public record OpenedStoreAdministrationTopFiveInfo(
         String administrationCode,
         String administrationCodeName,
-        Double curOpenedRate
+        Long curOpenedRate
 ) {
 }
